Update fopen handling and unit tests
Now all unit tests for PushFile are in place and working. Unfortunately, it cannot be tested past the point where `curl_exec` returns an error, as that would require a server and external calls.

As for `fopen()`, it has very rudimentary error handling, so I combined it with error suppression and error_get_last().