Add error handling to `PushFeed` (#25)
* Update the cURL command and error handling

* Initial and partial tests

* Update fopen handling and unit tests

Now all unit tests for PushFile are in place and working. Unfortunately, it cannot be tested past the point where `curl_exec` returns an error, as that would require a server and external calls.

As for `fopen()`, it has very rudimentary error handling, so I combined it with error suppression and error_get_last().

* Remove the demo test

* Move the ScheduledActionManagerTest to the right namespace in the tests folder

* Clean up the AS action and CLI command error handling

If anything goes wrong, it will trigger an exception on the PushFile level.

* Remove redundant  check

* Address CR nitpicks

* Use a generic Exception instead of RuntimeException